+++ CHANGELOG.md
@@ -23,6 +23,7 @@
 - Fixed background image stretching to infinity if the page is long enough
 - Fixed Stop column display for multi-timetable posts (Sosnowiec Główny)
 - Fixed missing pictures for some trains due to the game update on 2023-04-07
+- Fixed train notification icon appearing for offline trains
 
 ## *2.0 - 2023-03-27*